refactor: hoist stub chain sync data to util-dev
hoist a few types from golden-test-generator to util-dev
in order to avoid circular dependencies
