+++ yarn-project.nix
@@ -1761,7 +1761,7 @@ cacheEntries = {
 "lru-cache@npm:4.1.5" = { filename = "lru-cache-npm-4.1.5-ede304cc43-4bb4b58a36.zip"; sha512 = "4bb4b58a36cd7dc4dcec74cbe6a8f766a38b7426f1ff59d4cf7d82a2aa9b9565cd1cb98f6ff60ce5cd174524868d7bc9b7b1c294371851356066ca9ac4cf135a"; };
 "lru-cache@npm:5.1.1" = { filename = "lru-cache-npm-5.1.1-f475882a51-c154ae1cbb.zip"; sha512 = "c154ae1cbb0c2206d1501a0e94df349653c92c8cbb25236d7e85190bcaf4567a03ac6eb43166fabfa36fd35623694da7233e88d9601fbf411a9a481d85dbd2cb"; };
 "lru-cache@npm:6.0.0" = { filename = "lru-cache-npm-6.0.0-b4c8668fe1-f97f499f89.zip"; sha512 = "f97f499f898f23e4585742138a22f22526254fdba6d75d41a1c2526b3b6cc5747ef59c5612ba7375f42aca4f8461950e925ba08c991ead0651b4918b7c978297"; };
-"lru-cache@npm:7.14.0" = { filename = "lru-cache-npm-7.14.0-354cf654ec-efdd329f2c.zip"; sha512 = "efdd329f2c1bb790b71d497c6c59272e6bc2d7dd060ba55fc136becd3dd31fc8346edb446275504d94cb60d3c8385dbf5267b79b23789e409b2bdf302d13f0d7"; };
+"lru-cache@npm:7.18.1" = { filename = "lru-cache-npm-7.18.1-1ae5fcb8ef-ab0ec1360c.zip"; sha512 = "ab0ec1360c552f1ffa54b1eaf0026126c5116a07bee156b92d1e971f4c8c88e9160f0fad4ab6fed4e0fdea84f25a4590ece085bc57ed9ab1d90b17f0b138c556"; };
 "ltgt@npm:2.2.1" = { filename = "ltgt-npm-2.2.1-443b5da86d-7e3874296f.zip"; sha512 = "7e3874296f7538bc8087b428ac4208008d7b76916354b34a08818ca7c83958c1df10ec427eeeaad895f6b81e41e24745b18d30f89abcc21d228b94f6961d50a2"; };
 "lunr@npm:2.3.9" = { filename = "lunr-npm-2.3.9-fa3aa9c2d6-176719e24f.zip"; sha512 = "176719e24fcce7d3cf1baccce9dd5633cd8bdc1f41ebe6a180112e5ee99d80373fe2454f5d4624d437e5a8319698ca6837b9950566e15d2cae5f2a543a3db4b8"; };
 "madge@npm:5.0.1" = { filename = "madge-npm-5.0.1-9f3b536039-821db88fdc.zip"; sha512 = "821db88fdcb6259cf4bf1665abb976dd859fb7157a712886468d4172ab6d758f3446212fcd3fa92be21e9e2b408854f96e9495b7119d7387e6fe5690bd9786c4"; };
@@ -1820,7 +1820,7 @@ cacheEntries = {
 "minipass-pipeline@npm:1.2.4" = { filename = "minipass-pipeline-npm-1.2.4-5924cb077f-b14240dac0.zip"; sha512 = "b14240dac0d29823c3d5911c286069e36d0b81173d7bdf07a7e4a91ecdef92cdff4baaf31ea3746f1c61e0957f652e641223970870e2353593f382112257971b"; };
 "minipass-sized@npm:1.0.3" = { filename = "minipass-sized-npm-1.0.3-306d86f432-79076749fc.zip"; sha512 = "79076749fcacf21b5d16dd596d32c3b6bf4d6e62abb43868fac21674078505c8b15eaca4e47ed844985a4514854f917d78f588fcd029693709417d8f98b2bd60"; };
 "minipass@npm:3.3.5" = { filename = "minipass-npm-3.3.5-a555b091e7-f89f02bcaa.zip"; sha512 = "f89f02bcaa0e0e4bb4c44ec796008e69fbca62db0aba6ead1bc57d25bdaefdf42102130f4f9ecb7d9c6b6cd35ff7b0c7b97d001d3435da8e629fb68af3aea57e"; };
-"minipass@npm:4.0.0" = { filename = "minipass-npm-4.0.0-2520d5ae3c-7a609afbf3.zip"; sha512 = "7a609afbf394abfcf9c48e6c90226f471676c8f2a67f07f6838871afb03215ede431d1433feffe1b855455bcb13ef0eb89162841b9796109d6fed8d89790f381"; };
+"minipass@npm:4.2.4" = { filename = "minipass-npm-4.2.4-20f938f2dd-c664f2ae44.zip"; sha512 = "c664f2ae4401408d1e7a6e4f50aca45f87b1b0634bc9261136df5c378e313e77355765f73f59c4a5abcadcdf43d83fcd3eb14e4a7cdcce8e36508e2290345753"; };
 "minizlib@npm:2.1.2" = { filename = "minizlib-npm-2.1.2-ea89cd0cfb-f1fdeac0b0.zip"; sha512 = "f1fdeac0b07cf8f30fcf12f4b586795b97be856edea22b5e9072707be51fc95d41487faec3f265b42973a304fe3a64acd91a44a3826a963e37b37bafde0212c3"; };
 "mixpanel@npm:0.13.0" = { filename = "mixpanel-npm-0.13.0-da020f2419-b7486d763a.zip"; sha512 = "b7486d763aa8f57670e99590db0488bc722178b8844c0508d127c8f1b160fd36053e3e90f0cc11d0669dfd7778b170b3638200b481005c6583122ae343c5f069"; };
 "mkdirp-classic@npm:0.5.3" = { filename = "mkdirp-classic-npm-0.5.3-3b5c991910-3f4e088208.zip"; sha512 = "3f4e088208270bbcc148d53b73e9a5bd9eef05ad2cbf3b3d0ff8795278d50dd1d11a8ef1875ff5aea3fa888931f95bfcb2ad5b7c1061cfefd6284d199e6776ac"; };
